Update keyword highlighting system and refine job ranking logic
- Modify the `.cursorignore` file to allow tracking of `.env.example` while ignoring `.env` for better environment management.
- Enhance the README with a detailed overview of LLM tasks, including job ranking, form field mapping, rejection analysis, answer synthesis, and headline generation, improving documentation clarity.
- Implement a three-tier keyword highlighting system in the dashboard, categorizing keywords into green (Microsoft ecosystem + Terraform), yellow (acceptable cloud), and red (prohibitive/non-Microsoft), enhancing job description analysis.
- Update the `ranker.ts` to recalculate fit scores based on category scores using current weights, ensuring accuracy in job ranking.
- Refactor UI components to display new keyword categories in the job description panel, improving user experience.

These changes enhance the job ranking process, improve documentation, and provide a more robust keyword highlighting system for job descriptions.

+++ src/lib/db.ts
@@ -28,7 +28,7 @@ export function initDb(): void {
       company TEXT NOT NULL,
       url TEXT UNIQUE NOT NULL,
       easy_apply INTEGER NOT NULL,
-      rank INTEGER,
+      rank REAL,
       status TEXT DEFAULT 'queued',
       applied_method TEXT,
       rejection_reason TEXT,
@@ -194,7 +194,7 @@ export interface Job {
   company: string;
   url: string;
   easy_apply: boolean;
-  rank?: number;
+  rank?: number; // Decimal value with full precision (e.g., 87.35)
   status: 'queued' | 'applied' | 'interview' | 'rejected' | 'skipped' | 'reported';
   applied_method?: 'automatic' | 'manual';
   rejection_reason?: string;
